Note on redistribution of SQLite source files
As stated by the MIT License, you are whelcome to reuse, modify, and redistribute the SQLiteCpp source code
the way you whant it to, be it a git submodule, a subdirectory, or a selection of some source files.

I whould love a mention in your README, a web link to the SQLite repository, and a mention of the author,
but none of those are mandatory.