Fix std::filesystem::path Database constructor
The return value of path.c_str() is platform-dependent and is a wstring
on windows. path.u8string() always returns an utf8-encoded string.
In c++17, that string has type std::string, but starting in c++20 it is
std::u8string. We use a reinterpret_cast to handle the c++20 case.